+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this project will be documented in this file. Dates are d
 
 #### [Unreleased](https://github.com/hougesen/mdsf/compare/v0.8.2...HEAD)
 
-- feat(tools): add support for pyupgrade [`#877`](https://github.com/hougesen/mdsf/pull/877)
+- feat(tools): add support for pyflakes [`#879`](https://github.com/hougesen/mdsf/pull/879)
+- feat(tools): add support for pyupgrade [`#878`](https://github.com/hougesen/mdsf/pull/878)
 - feat(tools): add support for pydocstyle [`#876`](https://github.com/hougesen/mdsf/pull/876)
 - feat(tools): add support for pydoclint [`#875`](https://github.com/hougesen/mdsf/pull/875)
 - feat(tools): add support for eradicate [`#874`](https://github.com/hougesen/mdsf/pull/874)